refactor(runtime): split numeric conversions
- [x] Built on macOS/Linux (Debug).
- [x] ctest passed.
- [ ] Added/updated tests.
- [x] Code formatted; no new warnings.
- [x] No spec/API deviation (or ADR attached).
- [ ] Docs updated if behavior/API changed.
- [x] Updated docs/comments per policy.
- [ ] Added/updated file headers.
